#36 feat * refacting register
