+++ src/types/Registration.tsx
@@ -1,21 +1,5 @@
 export interface Registration {
-  id?: number;
-  country?: string;
-  city?: string;
-  firstName?: string;
-  lastName?: string;
   pseudo: string;
   email: string;
-  walletAddress: string;
-  contribution?: string;
-  socialMedias?: {
-    twitter?: string,
-    discord?: string,
-    tiktok?: string,
-    telegram?: string
-  };
-  referral?: string;
-  aboutUs?: string;
-  confidentiality: boolean;
-  beContacted?: boolean;
+  password: string;
 }
