+++ src/types/SocialMedia.tsx
@@ -1,0 +1,6 @@
+export interface SocialMedias {
+	twitter?: string,
+	discord?: string,
+	tiktok?: string,
+	telegram?: string
+  }
