+++ src/types/User.tsx
@@ -1,0 +1,18 @@
+import { SocialMedias } from "./SocialMedia";
+
+export interface User {
+	id?: number;
+	country?: string;
+	city?: string;
+	firstName?: string;
+	lastName?: string;
+	pseudo: string;
+	email: string;
+	walletAddress: string;
+	contribution?: string;
+	socialMedias?: SocialMedias;
+	referral?: string;
+	aboutUs?: string;
+	confidentiality: boolean;
+	beContacted?: boolean;
+  }
